refactor: extract AggregateByDomain helpers into aggregate.go
Move AggregateByDomain* and related private helpers out of service.go
to shrink the god-file and lower its architecture LOC ceiling.
